Added 2 IR receivers on 2 pins
I changed the code to allow it to look for a tag on two different pins.
It will look for a tag on one pin first, then the next pin. Hopefully, I
can see a tag on either pin. this will allow me to have multiple
receivers and pins I can look at.