Now using hardware Timer
I changed the code to use a hardware timer instead of Adafruit’s
digital Write code because the arduino Uno did not send the same length
of code as the decimeuluevo. I also should pick and stick with a
protocol duration. Probs 600. I like the numbers better (50’s are
annoying to add).